Implemented a dedicated class for explicit format processing, allowing invocation outside of the formatter functionality.
